Move category endpoint from fair-events-experimental to fair-events
The Manage Event admin page (base plugin) creates categories on the
fly via /fair-events/v1/sources/categories, but the endpoint only
existed in fair-events-experimental behind the sources feature flag,
so it 404'd for plain fair-events installs.
